move mfp scripts
