+++ src/routes/user.js
@@ -5,7 +5,7 @@ import Moment from 'moment-timezone';
 import { extendMoment } from 'moment-range';
 
 import ensureAuthenticated from '../methods/auth';
-import getMfpData from '../methods/mfp';
+import getMfpData from '../myfitnesspal/mfp';
 
 import Entry from '../models/entry';
 import logger from '../methods/logger';
